+++ inc/template-functions.php
@@ -42,15 +42,7 @@ function bootscore_pingback_header() {
 add_action('wp_head', 'bootscore_pingback_header');
 
 
-/**
- * Allow modifying the default bootstrap container class
- * @return string
- */
-if (!function_exists('bootscore_container_class')) {
-  function bootscore_container_class() {
-    return "container";
-  }
-}
+
 
 
 /**
